+++ Gruntfile.js
@@ -39,7 +39,6 @@ module.exports = function (grunt) {
         command: 'npm publish --dry-run'
       },
       pretest: {
-        // TODO finish jsdocs and uncomment
         // command: 'grunt jsdoc && ' +
         command: 'babel "<%= __cwd %>/lib" --out-dir "<%= __cwd %>/dist" --minified --compact=true --no-comments && ' +
           'browserify -e "<%= __cwd %>/dist/racepacks.js" -o "<%= __cwd %>/dist/racepacks.bundle.js" -s' +