+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 # racepacks Changelog
 
+## v0.2.9 Improvements and corrections
+Improve README.md.
+
+Improve/add examples.
+
+Correct redirectStream 'finish' error at lib/racepacks.js causing a tmp file not being deleted even if empty.
+
+Improve lib/racepacks.js with 'silent' option, constructor logic, 0 total test time handling, performance.now
+ rounding, results 'podium' array.
+ 
+Improve scripts/jsdoc-conf.js exclude.
+
+
 ## v0.1.0 Initial release